created separate component directory moved MyInfo components and edited import in index.js file to properly import moved component file
