+++ src/index.js
@@ -1,6 +1,6 @@
 import React from "react"
 import ReactDOM from "react-dom"
 
-import MyInfo from "./MyInfo"
+import MyInfo from "./components/MyInfo"
 
 ReactDOM.render(<MyInfo />, document.getElementById("root"))
